+++ Simple Examples/Assets/Normal/Examples/Blank Scene Template/Skybox/GradientSkybox.shader
@@ -26,9 +26,9 @@
 			#include "UnityCG.cginc"
 			#include "GradientSkybox.cginc"
 
-			fixed3 _ColorT;
-			fixed3 _ColorM;
-			fixed3 _ColorB;
+			half3 _ColorT;
+			half3 _ColorM;
+			half3 _ColorB;
 			half _ExponentT;
 			half _ExponentB;
 			half _Intensity;
